ml/requirements.txt
ultralytics
roboflow
opencv-python
numpy
Pillow
